requirements.txt
django>=5.1
djangorestframework>=3.12
